Fix warning introduced a couple of commits ago.

+++ src/lib/util.cc
@@ -791,7 +791,7 @@ careful_string_filter (string s)
 	/* Then remove anything that's not in a very limited character set */
 	wstring out;
 	wstring const allowed = L"0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz-_%.+";
-	for (size_t i = 0; i < transliterated_more.length(); ++i) {
+	for (int i = 0; i < transliterated_more.length(); ++i) {
 		wchar_t c = transliterated_more[i];
 		if (allowed.find(c) != string::npos) {
 			out += c;
